Conditional execution IR, gate, kernel, program, writing qasm, and custom gate interfaces pretty stable.
Still to do: scheduler depgraph, mapper decomposition, check other passes, default gate interface for measure with/out breg

+++ src/scheduler.h
@@ -117,7 +117,7 @@ public:
 
     // factored out code from Init to add a dependence between two nodes
     // operand is in qubit_creg_breg combined index space with size qcount+ccount+bcount
-    void add_dep(int srcID, int tgtID, enum DepTypes deptype, int operand);
+    void add_dep(int srcID, int tgtID, enum DepTypes deptype, size_t operand);
 
     // fill the dependence graph ('graph') with nodes from the circuit and adding arcs for their dependences
     void init(
